Wrap code into standalone module
I made this a bit more modular to make it easier for working with ovito
and eventually a signac-flow operation.

This also serves as a reference point for some performance improvements
I am attempting to implement. For reference, the code takes ~10.4
seconds to run on one frame of a test trajectory.