Implement adjacency_matrix as a set up tuples
Since the adjacency matrix is so sparse, we can get big performance
improvements by storing it as a set of tuples, allowing fast inclusion
lookup. Direct benchmarks of `if adj_matrix[i][j] == 1` compared to `if
(i, j) in adj_matrix` show that the sparse matrix is ~15x faster. The
full calculation using the sparse matrix runs in ~0.4 seconds wall time
on the frame that took ~10.4 seconds wall time to run with the full
matrix, a ~26x speedup.