Modify pore loading analysis for competitive capture
This commit changes the pore loading analysis operation to handle multiple guest types (i.e., for the competitive capture simulations. It is also changes how that data is stored in the job documents. Now, the only data that is stored in the job doc is the total number of pores, and the number catpured guests for each guest species. It no longer stores the fractional loading since that is easily calculated on the fly during any analysis. I therefore had to rerun the analysis on the noncompetitive capture simulations. The current implementation can handle both the competitive and noncompetitive capture simulaions, since it uses the guest_info statepoint key to determine how many of each type of guest was captured.